CleanUpFunctions/requirements.txt
azure-functions
azure-identity
azure-storage-blob
pyodbc
